A bit of refacto to have the schedules api fully contained in internal/schedules
This is the first step for adding test on the api itself.
The main take care of building dependencies and to add sugar around the
api, like status, metrics and other things.
